+++ src/rapidfuzz/__init__.py
@@ -5,7 +5,13 @@ from __future__ import annotations
 
 __author__: str = "Max Bachmann"
 __license__: str = "MIT"
-__version__: str = "3.4.0"
+
+import importlib.metadata as importlib_metadata
+
+try:
+    __version__ = importlib_metadata.version(__package__ or __name__)
+except importlib_metadata.PackageNotFoundError:
+    __version__ = "0.0.0"
 
 from rapidfuzz import distance, fuzz, process, utils