+++ CHANGELOG.rst
@@ -11,7 +11,8 @@ Changed
 
 Performance
 ~~~~~~~~~~~
-* improve performance of simd implementation for ``Indel`` / ``Jaro`` / ``JaroWinkler``
+* improve performance of simd implementation for ``LCS`` / ``Indel`` / ``Jaro`` / ``JaroWinkler``
+* improve performance of Jaro and Jaro Winkler for long sequences
 
 [3.4.0] - 2023-10-09
 ^^^^^^^^^^^^^^^^^^^^